requirements.txt
python-telegram-bot-api
telebot
flask
python-dotenv
psycopg2-binary
requests
schedule
gunicorn
gevent